requirements_dev.txt
-r requirements.txt

albumentations==1.4.6
clearml==1.11.1
hydra-core==1.3.2
lightning==2.2.3
omegaconf==2.3.0
pandas==2.2.2
pre-commit==3.7.0
pytorch-lightning==2.2.3
scikit_multilearn==0.2.0
tensorboard==2.16.2
timm==0.9.2
torchmetrics==1.0.1
